configure: Correct the use of AC_ARG_ENABLE() macro
Fix the use of AC_ARG_ENABLE() so that --disable-foo configure options
are respected.

The third and fourth arguments of AC_ARG_ENABLE are respectively
ACTION-IF-GIVEN and ACTION-IF-NOT-GIVEN. The former is executed if
either --enable-foo or --disable-foo is specified on command line, the
latter if neither is. Additionally, the specified value is stored in
$enable_foo variable.

Therefore, the common practice is to provide null ACTION-IF-GIVEN
(and use the value of $enable_foo), and to set a default for $enable_foo
in ACTION-IF-NOT-GIVEN. The commit does that, and adjusts the further
conditionals to test $enable_foo properly.