Remove GOT_PREFIX, it's a broken hack of limited utility
To add flags to the compiler, CFLAGS, CPPFLAGS or LDFLAGS can be used on
the configure command line.

The need to add flags depends on the location of the headers and
libraries of the dependencies, which is orthogonal to the directory where
xrdp will be installed.

The implementation in configure.ac has a stray closing bracket, making
GOT_PREFIX true even if --prefix is not passed.

The implementation is inconsistent - the only affected makefiles are for
xrdp and libxrdp.

Changing rpath manually is wrong in most cases. Libtool should be able to
set rpath correctly on its own.

Using $(prefix)/lib ignores the libdir setting. For many 64-bit systems,
/usr/lib is used for 32-bit libraries. Adding 32-bit libraries to the
rpath slows down 64-bit executables, as the dynamic loader searches for
libraries in a wrong directory.

There is no way to disable GOT_PREFIX if --prefix has to be passed.
Fedora RPM patches configure.ac and needs to rerun autoconf and automake
after that.