Use special name "." for the branches in .gitmodules
It is used to indicate that the submodule branch should be the same as
the branch of the parent repository - devel for devel, master for master.

It's a new feature of git, appearing in version 2.10, but most users
should not really care. It only affects the behavior of "git submodule
update --remote" when updating the submodules.