Use both config and configdir to avoid any defaults
If the -config switch is left out, the server will attempt to load
settings from the default file /etc/X11/xorg.conf. This file is now
usually non-existent; most distributions preferring to have the
configuration through files in /etc/X11/xorg.conf.d instead.

But in order to prevent any legacy settings that are put there to
interfere with XRDP, _both_ config and configdir should be redirected
to the xrdp/ subdirectory. This will overspecify the configuration by
having the xrdp/xorg.conf file probably read twice; this does not seem
to cause any harm. Incidentally, it does seem to make the X server now
pick up xrdpdev as a GPUDevice(?).