Fix #1616 allowing clipboard to work both ways
Currently the clipboard copy and paste is working to copy from a the server running xrdp and vncserver and paste to a client running Windows Remote Desktop, but it does not work to copy data from the Windows client and paste to the server running xdrp / vncserver. The code that requests the data from the client clipboard after an announcement / notification that new data is available appears to have been disabled.