Use 'struct peer' for connection information
At present, the connected peer is passed around in a text string which
contains the peer IP, peer port and the client file descriptor. This
makes it hard for security code to extract the IP address in particular
- a special function g_get_ip_from_description() is used for this which
is a bit of a hack.

This commit replaces the connection_description text string with a
'struct peer' which allows the peer connection information fields to
be accessed directly.

One change in functionality is that when a client reconnects, the
peer details can be updated.