Remove unnecesssary data from struct auth_info
When using PAM authentication, a copy is made of the username and
password in the auth_info structure.

The password copy is not cleared from memory when the structure is
deallocated. This could mean the password is revealed to an attacker
from a coredump.

One solution is to clear the password when the struct is
deallocated. However, the username and password in the auth_info struct
are only required for the duration of the PAM conversation function. A
better solution is to remove the username and password from the auth_info
struct entirely, and just use pointers for the duration of the time the
callback function is used.