Allow a zero timeout for g_obj_wait()
At the moment, timeouts of -1 or 0 for g_obj_wait() both mean
'no timeout', i.e. wait forever. This commit changes a timeout of
0 to mean 'return immediately'. This is more consistent with poll()
and WaitForMultipleObjects(), and can be used by the new encoder
logic.

There are a very few places on xrdp where a timeout of 0 is currently
specified to mean 'wait forever'. These are replaced with values of
-1.