Add Latvian keyboard definitions
The Latvian keyboard corresponding to code 0x426 is totally different
from the keyboard for 0x10426 (Latvian(QWERTY)) and 0x20426
(Latvian(Standard)). We set up new definitions for 0x426 and 0x10426 and
symlink 0x20426 to 0x10426.

The ini files are generated on Ubuntu 20.04 to avoid a non-working
caps-lock key. See this PR for details:-

https://github.com/neutrinolabs/xrdp/pull/3022

This script was used to generate the files on a clean X server:-

    #!/bin/sh

    if [ -z "$1" ]; then
        xvfb-run -s "-noreset" $0 _IN_XVFB_
    else
        setxkbmap -model pc104 -layout lv
        ./xrdp-genkeymap ../instfiles/km-00010426.ini
        setxkbmap -model pc104 -layout lv -variant ergonomic
        ./xrdp-genkeymap ../instfiles/km-00000426.ini
    fi